+++ CHANGELOG.md
@@ -5,5 +5,7 @@ Notable changes to the project are (hopefully) documented in this file.
 Attempts are made to confirm to semantic versioning.
 
 ## [Unreleased]
-- Added return error for ambiguous chains
-- Added code system support
+### Added
+- Return error for ambiguous chains
+- Code system support
+- Comma-separated values (OR values)